+++ lib/utils/interrupt_char.h
@@ -28,6 +28,5 @@
 
 extern int mp_interrupt_char;
 void mp_hal_set_interrupt_char(int c);
-void mp_keyboard_interrupt(void);
 
 #endif // MICROPY_INCLUDED_LIB_UTILS_INTERRUPT_CHAR_H
