+++ ports/cc3200/telnet/telnet.c
@@ -26,8 +26,7 @@
 
 #include <stdint.h>
 
-#include "py/mpconfig.h"
-#include "py/obj.h"
+#include "py/runtime.h"
 #include "py/mphal.h"
 #include "lib/utils/interrupt_char.h"
 #include "telnet.h"
