+++ ports/samd/mphalport.c
@@ -24,9 +24,8 @@
  * THE SOFTWARE.
  */
 
-#include "py/mpstate.h"
+#include "py/runtime.h"
 #include "py/mphal.h"
-#include "lib/utils/interrupt_char.h"
 #include "samd_soc.h"
 #include "tusb.h"
 
@@ -44,15 +43,6 @@ void mp_hal_set_interrupt_char(int c) {
         mp_obj_exception_clear_traceback(MP_OBJ_FROM_PTR(&MP_STATE_VM(mp_kbd_exception)));
     }
     tud_cdc_set_wanted_char(c);
-}
-
-void mp_keyboard_interrupt(void) {
-    MP_STATE_VM(mp_pending_exception) = MP_OBJ_FROM_PTR(&MP_STATE_VM(mp_kbd_exception));
-    #if MICROPY_ENABLE_SCHEDULER
-    if (MP_STATE_VM(sched_state) == MP_SCHED_IDLE) {
-        MP_STATE_VM(sched_state) = MP_SCHED_PENDING;
-    }
-    #endif
 }
 
 #endif
